requirements.txt
anyio==3.6.1
asgiref==3.5.2
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
Jinja2==3.1.2
MarkupSafe==2.1.1
motor==3.0.0
pydantic==1.9.1
PyJWT==2.4.0
pymongo==4.1.1
python-decouple==3.6
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6
zipp==3.8.0
